+++ src/Query/QueryBuilder.php
@@ -508,8 +508,8 @@ abstract class QueryBuilder implements QueryBuilderInterface
      * Builds given $expression.
      *
      * @param ExpressionInterface $expression the expression to be built
-     * @param array $params the parameters to be bound to the generated SQL statement. These parameters will be included
-     * in the result with the additional parameters generated during the expression building process.
+     * @param array $params the parameters to be bound to the generated SQL statement. These parameters will be
+     * included in the result with the additional parameters generated during the expression building process.
      *
      * @throws Exception|InvalidArgumentException|InvalidConfigException|NotSupportedException when $expression building
      * is not supported by this QueryBuilder.
@@ -1259,39 +1259,6 @@ abstract class QueryBuilder implements QueryBuilderInterface
     }
 
     /**
-     * Creates an INSERT SQL statement.
-     *
-     * For example,
-     *
-     * ```php
-     * $sql = $queryBuilder->insert('user', [
-     *     'name' => 'Sam',
-     *     'age' => 30,
-     * ], $params);
-     * ```
-     *
-     * The method will properly escape the table and column names.
-     *
-     * @param string $table the table that new rows will be inserted into.
-     * @param array|Query $columns the column data (name => value) to be inserted into the table or instance of
-     * {@see Query} to perform INSERT INTO ... SELECT SQL statement. Passing of {@see Query}.
-     * @param array $params the binding parameters that will be generated by this method. They should be bound to the
-     * DB command later.
-     *
-     * @throws Exception|InvalidArgumentException|InvalidConfigException|NotSupportedException
-     *
-     * @return string the INSERT SQL.
-     */
-    public function insert(string $table, Query|array $columns, array &$params = []): string
-    {
-        [$names, $placeholders, $values, $params] = $this->prepareInsertValues($table, $columns, $params);
-
-        return 'INSERT INTO ' . $this->quoter->quoteTableName($table)
-            . (!empty($names) ? ' (' . implode(', ', $names) . ')' : '')
-            . (!empty($placeholders) ? ' VALUES (' . implode(', ', $placeholders) . ')' : $values);
-    }
-
-    /**
      * Builds a SQL statement for renaming a column.
      *
      * @param string $table the table whose column is to be renamed. The name will be properly quoted by the method.
@@ -1381,44 +1348,6 @@ abstract class QueryBuilder implements QueryBuilderInterface
     public function truncateTable(string $table): string
     {
         return 'TRUNCATE TABLE ' . $this->quoter->quoteTableName($table);
-    }
-
-    /**
-     * Creates an UPDATE SQL statement.
-     *
-     * For example,
-     *
-     * ```php
-     * $params = [];
-     * $sql = $queryBuilder->update('user', ['status' => 1], 'age > 30', $params);
-     * ```
-     *
-     * The method will properly escape the table and column names.
-     *
-     * @param string $table the table to be updated.
-     * @param array $columns the column data (name => value) to be updated.
-     * @param array|string $condition the condition that will be put in the WHERE part. Please refer to
-     * {@see Query::where()} on how to specify condition.
-     * @param array $params the binding parameters that will be modified by this method so that they can be bound to the
-     * DB command later.
-     *
-     * @psalm-param array<string, ExpressionInterface|string> $columns
-     *
-     * @throws Exception|InvalidArgumentException
-     *
-     * @return string the UPDATE SQL.
-     */
-    public function update(string $table, array $columns, array|string $condition, array &$params = []): string
-    {
-        /**
-         * @psalm-var array $lines
-         * @psalm-var array $params
-         */
-        [$lines, $params] = $this->prepareUpdateSets($table, $columns, $params);
-        $sql = 'UPDATE ' . $this->quoter->quoteTableName($table) . ' SET ' . implode(', ', $lines);
-        $where = $this->buildWhere($condition, $params);
-
-        return ($where === '') ? $sql : ($sql . ' ' . $where);
     }
 
     /**
@@ -1608,7 +1537,7 @@ abstract class QueryBuilder implements QueryBuilderInterface
      *
      * @return array array of column names, placeholders, values and params.
      */
-    protected function prepareInsertValues(string $table, Query|array $columns, array $params = []): array
+    public function prepareInsertValues(string $table, Query|array $columns, array $params = []): array
     {
         $tableSchema = $this->schema->getTableSchema($table);
         $columnSchemas = $tableSchema !== null ? $tableSchema->getColumns() : [];
@@ -1652,7 +1581,7 @@ abstract class QueryBuilder implements QueryBuilderInterface
      * @return array `SET` parts for an `UPDATE` SQL statement (the first array element) and params (the second array
      * element).
      */
-    protected function prepareUpdateSets(string $table, array $columns, array $params = []): array
+    public function prepareUpdateSets(string $table, array $columns, array $params = []): array
     {
         $tableSchema = $this->schema->getTableSchema($table);
 
@@ -1686,7 +1615,7 @@ abstract class QueryBuilder implements QueryBuilderInterface
      *@throws Exception|JsonException
      *
      */
-    protected function prepareUpsertColumns(
+    public function prepareUpsertColumns(
         string $table,
         Query|array $insertColumns,
         Query|bool|array $updateColumns,
