+++ src/TestSupport/TestQueryBuilderTrait.php
@@ -2371,327 +2371,4 @@ trait TestQueryBuilderTrait
             ],
         ];
     }
-
-    public function insertProviderTrait(): array
-    {
-        return [
-            'regular-values' => [
-                'customer',
-                [
-                    'email' => 'test@example.com',
-                    'name' => 'silverfire',
-                    'address' => 'Kyiv {{city}}, Ukraine',
-                    'is_active' => false,
-                    'related_id' => null,
-                ],
-                [],
-                $this->replaceQuotes(
-                    'INSERT INTO [[customer]] ([[email]], [[name]], [[address]], [[is_active]], [[related_id]])'
-                    . ' VALUES (:qp0, :qp1, :qp2, :qp3, :qp4)'
-                ),
-                [
-                    ':qp0' => 'test@example.com',
-                    ':qp1' => 'silverfire',
-                    ':qp2' => 'Kyiv {{city}}, Ukraine',
-                    ':qp3' => false,
-                    ':qp4' => null,
-                ],
-            ],
-            'params-and-expressions' => [
-                '{{%type}}',
-                [
-                    '{{%type}}.[[related_id]]' => null,
-                    '[[time]]' => new Expression('now()'),
-                ],
-                [],
-                'INSERT INTO {{%type}} ({{%type}}.[[related_id]], [[time]]) VALUES (:qp0, now())',
-                [
-                    ':qp0' => null,
-                ],
-            ],
-            'carry passed params' => [
-                'customer',
-                [
-                    'email' => 'test@example.com',
-                    'name' => 'sergeymakinen',
-                    'address' => '{{city}}',
-                    'is_active' => false,
-                    'related_id' => null,
-                    'col' => new Expression('CONCAT(:phFoo, :phBar)', [':phFoo' => 'foo']),
-                ],
-                [':phBar' => 'bar'],
-                $this->replaceQuotes(
-                    'INSERT INTO [[customer]] ([[email]], [[name]], [[address]], [[is_active]], [[related_id]], [[col]])'
-                    . ' VALUES (:qp1, :qp2, :qp3, :qp4, :qp5, CONCAT(:phFoo, :phBar))'
-                ),
-                [
-                    ':phBar' => 'bar',
-                    ':qp1' => 'test@example.com',
-                    ':qp2' => 'sergeymakinen',
-                    ':qp3' => '{{city}}',
-                    ':qp4' => false,
-                    ':qp5' => null,
-                    ':phFoo' => 'foo',
-                ],
-            ],
-            'carry passed params (query)' => [
-                'customer',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email',
-                        'name',
-                        'address',
-                        'is_active',
-                        'related_id',
-                    ])
-                    ->from('customer')
-                    ->where([
-                        'email' => 'test@example.com',
-                        'name' => 'sergeymakinen',
-                        'address' => '{{city}}',
-                        'is_active' => false,
-                        'related_id' => null,
-                        'col' => new Expression('CONCAT(:phFoo, :phBar)', [':phFoo' => 'foo']),
-                    ]),
-                [':phBar' => 'bar'],
-                $this->replaceQuotes(
-                    'INSERT INTO [[customer]] ([[email]], [[name]], [[address]], [[is_active]], [[related_id]])'
-                    . ' SELECT [[email]], [[name]], [[address]], [[is_active]], [[related_id]] FROM [[customer]]'
-                    . ' WHERE ([[email]]=:qp1) AND ([[name]]=:qp2) AND ([[address]]=:qp3) AND ([[is_active]]=:qp4)'
-                    . ' AND ([[related_id]] IS NULL) AND ([[col]]=CONCAT(:phFoo, :phBar))'
-                ),
-                [
-                    ':phBar' => 'bar',
-                    ':qp1' => 'test@example.com',
-                    ':qp2' => 'sergeymakinen',
-                    ':qp3' => '{{city}}',
-                    ':qp4' => false,
-                    ':phFoo' => 'foo',
-                ],
-            ],
-        ];
-    }
-
-    public function updateProviderTrait(): array
-    {
-        return [
-            [
-                'customer',
-                [
-                    'status' => 1,
-                    'updated_at' => new Expression('now()'),
-                ],
-                [
-                    'id' => 100,
-                ],
-                $this->replaceQuotes(
-                    'UPDATE [[customer]] SET [[status]]=:qp0, [[updated_at]]=now() WHERE [[id]]=:qp1'
-                ),
-                [
-                    ':qp0' => 1,
-                    ':qp1' => 100,
-                ],
-            ],
-        ];
-    }
-
-    public function upsertProviderTrait(): array
-    {
-        return [
-            'regular values' => [
-                'T_upsert',
-                [
-                    'email' => 'test@example.com',
-                    'address' => 'bar {{city}}',
-                    'status' => 1,
-                    'profile_id' => null,
-                ],
-                true,
-                null,
-                [
-                    ':qp0' => 'test@example.com',
-                    ':qp1' => 'bar {{city}}',
-                    ':qp2' => 1,
-                    ':qp3' => null,
-                ],
-            ],
-            'regular values with update part' => [
-                'T_upsert',
-                [
-                    'email' => 'test@example.com',
-                    'address' => 'bar {{city}}',
-                    'status' => 1,
-                    'profile_id' => null,
-                ],
-                [
-                    'address' => 'foo {{city}}',
-                    'status' => 2,
-                    'orders' => new Expression('T_upsert.orders + 1'),
-                ],
-                null,
-                [
-                    ':qp0' => 'test@example.com',
-                    ':qp1' => 'bar {{city}}',
-                    ':qp2' => 1,
-                    ':qp3' => null,
-                    ':qp4' => 'foo {{city}}',
-                    ':qp5' => 2,
-                ],
-            ],
-            'regular values without update part' => [
-                'T_upsert',
-                [
-                    'email' => 'test@example.com',
-                    'address' => 'bar {{city}}',
-                    'status' => 1,
-                    'profile_id' => null,
-                ],
-                false,
-                null,
-                [
-                    ':qp0' => 'test@example.com',
-                    ':qp1' => 'bar {{city}}',
-                    ':qp2' => 1,
-                    ':qp3' => null,
-                ],
-            ],
-            'query' => [
-                'T_upsert',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email',
-                        'status' => new Expression('2'),
-                    ])
-                    ->from('customer')
-                    ->where(['name' => 'user1'])
-                    ->limit(1),
-                true,
-                null,
-                [
-                    ':qp0' => 'user1',
-                ],
-            ],
-            'query with update part' => [
-                'T_upsert',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email',
-                        'status' => new Expression('2'),
-                    ])
-                    ->from('customer')
-                    ->where(['name' => 'user1'])
-                    ->limit(1),
-                [
-                    'address' => 'foo {{city}}',
-                    'status' => 2,
-                    'orders' => new Expression('T_upsert.orders + 1'),
-                ],
-                null,
-                [
-                    ':qp0' => 'user1',
-                    ':qp1' => 'foo {{city}}',
-                    ':qp2' => 2,
-                ],
-            ],
-            'query without update part' => [
-                'T_upsert',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email',
-                        'status' => new Expression('2'),
-                    ])
-                    ->from('customer')
-                    ->where(['name' => 'user1'])
-                    ->limit(1),
-                false,
-                null,
-                [
-                    ':qp0' => 'user1',
-                ],
-            ],
-            'values and expressions' => [
-                '{{%T_upsert}}',
-                [
-                    '{{%T_upsert}}.[[email]]' => 'dynamic@example.com',
-                    '[[ts]]' => new Expression('now()'),
-                ],
-                true,
-                null,
-                [
-                    ':qp0' => 'dynamic@example.com',
-                ],
-            ],
-            'values and expressions with update part' => [
-                '{{%T_upsert}}',
-                [
-                    '{{%T_upsert}}.[[email]]' => 'dynamic@example.com',
-                    '[[ts]]' => new Expression('now()'),
-                ],
-                [
-                    '[[orders]]' => new Expression('T_upsert.orders + 1'),
-                ],
-                null,
-                [
-                    ':qp0' => 'dynamic@example.com',
-                ],
-            ],
-            'values and expressions without update part' => [
-                '{{%T_upsert}}',
-                [
-                    '{{%T_upsert}}.[[email]]' => 'dynamic@example.com',
-                    '[[ts]]' => new Expression('now()'),
-                ],
-                false,
-                null,
-                [
-                    ':qp0' => 'dynamic@example.com',
-                ],
-            ],
-            'query, values and expressions with update part' => [
-                '{{%T_upsert}}',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email' => new Expression(':phEmail', [':phEmail' => 'dynamic@example.com']),
-                        '[[time]]' => new Expression('now()'),
-                    ]),
-                [
-                    'ts' => 0,
-                    '[[orders]]' => new Expression('T_upsert.orders + 1'),
-                ],
-                null,
-                [
-                    ':phEmail' => 'dynamic@example.com',
-                    ':qp1' => 0,
-                ],
-            ],
-            'query, values and expressions without update part' => [
-                '{{%T_upsert}}',
-                (new Query($this->getConnection()))
-                    ->select([
-                        'email' => new Expression(':phEmail', [':phEmail' => 'dynamic@example.com']),
-                        '[[time]]' => new Expression('now()'),
-                    ]),
-                [
-                    'ts' => 0,
-                    '[[orders]]' => new Expression('T_upsert.orders + 1'),
-                ],
-                null,
-                [
-                    ':phEmail' => 'dynamic@example.com',
-                    ':qp1' => 0,
-                ],
-            ],
-            'no columns to update' => [
-                'T_upsert_1',
-                [
-                    'a' => 1,
-                ],
-                false,
-                null,
-                [
-                    ':qp0' => 1,
-                ],
-            ],
-        ];
-    }
 }
